REFACTOR: Finialized changes for 1.0.0 beta 1
