+++ README.md
@@ -37,7 +37,7 @@ The purpose of this module is to **get connected** and **start streaming** as fa
 Get connected and start streaming right now
 
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ganglion = new Ganglion();
 ganglion.once('ganglionFound', (peripheral) => {
   // Stop searching for BLE devices once a ganglion is found.
@@ -103,14 +103,14 @@ Initialization
 Initializing the board:
 
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ganglion = new Ganglion();
 ```
 
 For initializing with options, such as verbose print outs:
 
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ourBoard = new Ganglion({
   verbose: true
 });
@@ -119,7 +119,7 @@ const ourBoard = new Ganglion({
 For initializing with callback, such as to catch errors on `noble` startup:
 
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ourBoard = new Ganglion((error) => {
   if (error) {
     console.log("error", error);  
@@ -131,7 +131,7 @@ const ourBoard = new Ganglion((error) => {
 For initializing with options and callback, such as verbose and to catch errors on `noble` startup:
 
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ourBoard = new Ganglion({
   verbose: true
 },(error) => {
@@ -151,7 +151,7 @@ You MUST wait for the 'ready' event to be emitted before streaming/talking with 
 so installing the 'sample' listener and writing before the ready event might result in... nothing at all.
 
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ourBoard = new Ganglion();
 ourBoard.connect(portName).then(function(boardSerial) {
     ourBoard.on('ready',function() {
@@ -178,7 +178,7 @@ To get a ['sample'](#event-sample) event, you need to:
 3. In callback for ['ready'](#event-ready) emitter, call [`streamStart()`](#method-stream-start)
 4. Install the ['sample'](#event-sample) event emitter
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ourBoard = new Ganglion();
 ourBoard.connect(localName).then(function() {
     ourBoard.on('ready',function() {
@@ -193,7 +193,7 @@ ourBoard.connect(localName).then(function() {
 ```
 Close the connection with [`.streamStop()`](#method-stream-stop) and disconnect with [`.disconnect()`](#method-disconnect)
 ```js
-const Ganglion = require('openbci-ganglion').Ganglion;
+const Ganglion = require('openbci-ganglion');
 const ourBoard = new Ganglion();
 ourBoard.streamStop().then(ourBoard.disconnect());
 ```
