+++ changelog.md
@@ -1,4 +1,21 @@
-# 0.4.4
+# v1.0.0
+
+Complete over haul of the core of the application to use the OpenBCI utilities NodeJS package.
+
+### Bug Fixes
+
+* Accel count value was wrong, was twice what it should have been.
+
+### Breaking Changes
+
+* Removed `index.js` and made entry point `openBCIGanglion.js` which leads to import of `const Ganglion = require('openbci-ganglion');` instead of `const Ganglion = require('openbci-ganglion').Ganglion`
+* `Constants` are now taken from new [OpenBCI Utilities Module](https://github.com/OpenBCI/OpenBCI_NodeJS_Utilities)
+
+## Beta 1
+
+Initial Beta Release
+
+# v0.4.4
 
 ### New Features
 * LSL streaming example (thanks @gabrielibagon)
@@ -6,12 +23,12 @@
 ### Bug Fixes
 * Fix #27 by adding delay
 
-# 0.4.3
+# v0.4.3
 
 ### Bug Fixes
 * Fix #23 by hard setting noble to 0.1.7, 0.1.8 breaks
 
-# 0.4.2
+# v0.4.2
 
 ### New Features
 * Add callback function to constructor to catch noble errors.
@@ -19,7 +36,7 @@
 ### Bug Fixes
 * Fix #19
 
-# 0.4.1
+# v0.4.1
 
 ### New Features
 * Dropped connection will fire disconnect
@@ -27,7 +44,7 @@
 ### Bug Fixes
 * Improvements to disconnect
 
-# 0.4.0
+# v0.4.0
 
 ### New Features
 * If dropped connection, module will now disconnect and clean up.
@@ -36,22 +53,22 @@
 * Dropped connection will fire disconnect
 * Auto reconnect functionality temporarily removed.
 
-# 0.3.8
+# v0.3.8
 
 ### New Features
 * Can now subscribe to `scanStart` and `scanStop` to better control scanning.
 
-# 0.3.7
+# v0.3.7
 
 ### Bug Fixes
 * Improvements to disconnect
 
-# 0.3.6
+# v0.3.6
 
 ### Bug Fixes
 * Fixed bug where not disconnecting.
 
-# 0.3.5 - 0.3.4 - 0.3.3
+# v0.3.5 - 0.3.4 - 0.3.3
 
 ### New Features
 * Get state of noble with `.isNobleReady()`
@@ -59,17 +76,17 @@
 ### Bug Fixes
 * Fix bug where scan stop would not stop scan which led to many problems.
 
-# 0.3.2
+# v0.3.2
 
 ### Bug Fixes
 * Fix bug where disconnect did not set private property `_connected` to false.
 
-# 0.3.1
+# v0.3.1
 
 ### Bug Fixes
 * Fix bug with connect
 
-# 0.3.0
+# v0.3.0
 
 ### New Features
 * Get accelerometer data from the ganglion! (Previous did not work while streaming)
@@ -81,7 +98,7 @@
 * Refactor file names for clarity
 * Removed dependency `underscore`
 
-# 0.2.0
+# v0.2.0
 
 ### Enhancements
 * Compress with 18bits vs 19bits
@@ -92,7 +109,7 @@
 * Fix bug where node process would not disconnect on windows.
 * Fix unhandled promise where server would call search start twice.
 
-# 0.1.1
+# v0.1.1
 
 ### New Features
 * Add function for starting, `.accelStart()`, and stopping, `.accelStop()`, accelerometer.
@@ -100,6 +117,6 @@
 ### Bug Fixes
 * Impedance was outputting on verbose instead of debug.
 
-# 0.1.0
+# v0.1.0
 
 Initial release
